Replace path traits with types
This commit replaces the traits `DatabasePath` and `DocumentPath` with
owning-types of the same name, and it adds two new traits, `IntoDatabasePath`
and `IntoDocumentPath`. Furthermore, it eliminates the specifier types; now all
path-related types are owning-types.

This commit is a work-in-progress for making the crate's API convenient and
type-safe, while keeping efficiency reasonable.