Transport rewrite
This commit resolves #11 and resolves #51. It also affects #32 and #56.

The new transport layer is smaller and more general. It has no CouchDB-specific
information. Also, it does away with the assumption that each Chill action
corresponds to exactly one HTTP request–response pair, and it should make it
easier to add support for asynchronous actions.